requirements.txt
python-dotenv==0.10.3
requests==2.22.0
urllib3==1.25.3
Pillow==6.0.0
dropbox==9.3.0
